+++ tests/db/rethinkdb/test_schema.py
@@ -4,16 +4,9 @@ import rethinkdb as r
 import bigchaindb
 from bigchaindb import backend
 from bigchaindb.backend.rethinkdb import schema
-from ..conftest import setup_database as _setup_database
 
 
-# Since we are testing database initialization and database drop,
-# we need to use the `setup_database` fixture on a function level
-@pytest.fixture(scope='function', autouse=True)
-def setup_database(request, node_config):
-    _setup_database(request, node_config)
-
-
+@pytest.mark.usefixtures('setup_database')
 def test_init_creates_db_tables_and_indexes():
     from bigchaindb.backend.schema import init_database
     conn = backend.connect()
@@ -35,6 +28,7 @@ def test_init_creates_db_tables_and_indexes():
         'assignee__transaction_timestamp')) is True
 
 
+@pytest.mark.usefixtures('setup_database')
 def test_init_database_fails_if_db_exists():
     from bigchaindb.backend.schema import init_database
     from bigchaindb.common import exceptions
@@ -52,14 +46,11 @@ def test_init_database_fails_if_db_exists():
 def test_create_database():
     conn = backend.connect()
     dbname = bigchaindb.config['database']['name']
-
-    # The db is set up by fixtures so we need to remove it
-    # and recreate it just with one table
-    conn.run(r.db_drop(dbname))
     schema.create_database(conn, dbname)
     assert conn.run(r.db_list().contains(dbname)) is True
 
 
+@pytest.mark.usefixtures('setup_database')
 def test_create_tables():
     conn = backend.connect()
     dbname = bigchaindb.config['database']['name']
@@ -76,6 +67,7 @@ def test_create_tables():
     assert len(conn.run(r.db(dbname).table_list())) == 3
 
 
+@pytest.mark.usefixtures('setup_database')
 def test_create_secondary_indexes():
     conn = backend.connect()
     dbname = bigchaindb.config['database']['name']
@@ -104,6 +96,7 @@ def test_create_secondary_indexes():
         'block_and_voter')) is True
 
 
+@pytest.mark.usefixtures('setup_database')
 def test_drop():
     conn = backend.connect()
     dbname = bigchaindb.config['database']['name']
@@ -115,6 +108,7 @@ def test_drop():
     assert conn.run(r.db_list().contains(dbname)) is False
 
 
+@pytest.mark.usefixtures('setup_database')
 def test_drop_non_existent_db_raises_an_error():
     from bigchaindb.common import exceptions
